Dialog to confirm delete.

+++ NotableNotes/src/pages/detail/detail.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { IonicPage, NavController, NavParams } from 'ionic-angular';
+import { IonicPage, NavController, NavParams, AlertController } from 'ionic-angular';
 import { NoteService } from '../../services/note.service';
 
 /**
@@ -20,14 +20,26 @@ export class DetailPage {
   constructor(
     public navCtrl: NavController, 
     public navParams: NavParams, 
-    private noteService: NoteService) {
-    this.note = this.navParams.get('note');
-    console.log('nav-param', this.note);
+    private noteService: NoteService,
+    private alertCtrl: AlertController) {
+      this.note = this.navParams.get('note');
+      console.log('nav-param', this.note);
   }
 
   public onDelete() {
-    this.noteService.deleteNote(this.note);
-    this.navCtrl.pop();
+    let confirm = this.alertCtrl.create({
+      title: 'Confirm delete',
+      message: 'Deleting a note is permanent. Are you sure you want to cotinue?',
+      buttons: [
+        {text: 'Cancel'},
+        {text: 'Delete', handler: () => {
+          this.noteService.deleteNote(this.note);
+          this.navCtrl.pop();
+        }}
+      ]
+    });
+    
+    confirm.present();
   }
 
   ionViewDidLoad() {
